+++ src/EnumUtils/enumUtils.hpp
@@ -1,79 +1,106 @@
 #ifndef ENUM_UTILS_HPP
 #define ENUM_UTILS_HPP
+#include "ppUtils.hpp"
 
-#define MAKE_ENUM(NAME,TYPE,...) \
-    enum class NAME : TYPE {__VA_ARGS__};
+#include <algorithm>
+#include <array>
+#include <string_view>
+using namespace std::string_view_literals;
 
-#include <functional>
-#include <initializer_list>
-#include <ostream>
-#include <string>
-#include <unordered_map>
-#include <vector>
+template < typename Type > struct IgnoreEquals {
+    Type value;
 
-template < typename Enum > struct ReflectiveEnum {
-    private:
-    static std::vector< Enum >                     sequence_values;
-    static std::unordered_map< Enum, std::string > enum_names;
-    static std::unordered_map< std::string, Enum > names_to_enums;
-    public:
-
-    ReflectiveEnum()= default;
-
-    // cppcheck-suppress noExplicitConstructor ; because initializer list
-    // should support var = {l1, l2, l3} syntax.
-    ReflectiveEnum(
-        std::initializer_list< std::pair< Enum, std::string > > t_enum_name_list
-    )
-        : ReflectiveEnum() {
-        for (const auto& enum_pair : t_enum_name_list) {
-            sequence_values.push_back(enum_pair.first);
-            enum_names.insert(enum_pair);
-            names_to_enums.insert({enum_pair.second, enum_pair.first});
-        }
+    template < typename Any >
+    constexpr IgnoreEquals& operator=([[maybe_unused]] Any t_val) {
+        return *this;
     }
 
-    Enum& operator[](const std::string& t_name) {
-        return names_to_enums[t_name];
+    constexpr explicit IgnoreEquals() : value() {}
+
+    constexpr explicit IgnoreEquals(Type t_value) : value(t_value) {}
+
+    // NOLINTNEXTLINE(hicpp-explicit-conversions)
+    constexpr operator Type() const { return value; }
+
+    constexpr IgnoreEquals operator*(const Type& t_value) {
+        return IgnoreEquals(t_value);
     }
-
-    const Enum& operator[](const std::string& t_name) const {
-        return names_to_enums.at(t_name);
-    }
-
-    std::string& operator[](const Enum& t_val) { return enum_names[t_val]; }
-
-    const std::string& operator[](const Enum& t_val) const {
-        return enum_names.at(t_val);
-    }
-
-    typename std::vector< Enum >::iterator begin() {
-        return sequence_values.begin();
-    }
-
-    typename std::vector< Enum >::const_iterator begin() const {
-        return sequence_values.begin();
-    }
-
-    typename std::vector< Enum >::const_iterator cbegin() const {
-        return sequence_values.cbegin();
-    }
-
-    typename std::vector< Enum >::iterator end() {
-        return sequence_values.end();
-    }
-
-    typename std::vector< Enum >::const_iterator end() const {
-        return sequence_values.end();
-    }
-
-    typename std::vector< Enum >::const_iterator cend() const {
-        return sequence_values.cend();
-    }
-
-    friend std::ostream& operator << (std::ostream& t_os, const Enum& t_e_val);
 };
 
-template <class Enum>
-std::ostream& operator << (std::ostream& t_os, const Enum& t_e_val);
+// NOLINTBEGIN(cppcoreguidelines-macro-usage)
+#define MAKE_ENUM(enum_name, enum_type, ...) \
+    enum enum_name : enum_type { __VA_ARGS__ };
+
+#define NAME_ENTRY(arg) #arg##sv.substr(0, #arg##sv.find_first_of('=') - 1)
+
+#define MAKE_NAMES(array_prefix, enum_name, ...)              \
+    constexpr static std::                                    \
+        array< std::string_view, MM_COUNT_ARGS(__VA_ARGS__) > \
+            array_prefix##_names{{MM_TRANSFORM(NAME_ENTRY, __VA_ARGS__)}};
+
+// NOLINTNEXTLINE(bugprone-macro-parentheses)
+#define VALUE_ENTRY(enum_name, arg) IgnoreEquals< enum_name >{} * arg
+#define MAKE_VALUES(array_prefix, enum_name, ...)                        \
+    constexpr static std::array< enum_name, MM_COUNT_ARGS(__VA_ARGS__) > \
+        array_prefix##_values{                                           \
+            {MM_TRANSFORM_1(VALUE_ENTRY, enum_name, __VA_ARGS__)}};
+
+#define BETTER_ENUM(enum_name, type, ...)                                      \
+    class enum_name {                                                          \
+      public:                                                                  \
+        MAKE_ENUM(m_enumeration, type, __VA_ARGS__)                            \
+      private:                                                                 \
+        MAKE_VALUES(m, m_enumeration, __VA_ARGS__)                             \
+        MAKE_NAMES(m, m_enumeration, __VA_ARGS__)                              \
+        constexpr static std::string_view m_name{#enum_name##sv};              \
+        m_enumeration                     m_value;                             \
+                                                                               \
+      public:                                                                  \
+        using value_container = decltype(m_values);                            \
+        using value_iterator  = typename value_container::iterator;            \
+        using value_type      = typename value_container::value_type;          \
+        using name_container  = decltype(m_names);                             \
+        using name_iterator   = typename name_container::iterator;             \
+        using name_type       = typename name_container::value_type;           \
+        using underlying      = type;                                          \
+                                                                               \
+        /* NOLINTNEXTLINE(hicpp-explicit-conversions)*/                        \
+        constexpr enum_name(const m_enumeration& t_val) : m_value(t_val) {}    \
+                                                                               \
+        friend constexpr enum_name operator+(const m_enumeration& t_val);      \
+                                                                               \
+        constexpr static value_container& values() { return m_values; }        \
+        constexpr static name_container&  names() { return m_names; }          \
+        constexpr static size_t size() { return MM_COUNT_ARGS(__VA_ARGS__); }  \
+        constexpr static std::string_view  name() { return m_name; }           \
+        constexpr static const value_type& from_string(std::string_view t_name \
+        ) {                                                                    \
+            const auto* found_ptr = std::find_if(                              \
+                m_names.begin(),                                               \
+                m_names.end(),                                                 \
+                [t_name](auto t_iter_name) { return t_iter_name == t_name; }   \
+            );                                                                 \
+            auto start_val = m_values.begin();                                 \
+            std::advance(                                                      \
+                start_val, std::distance(m_names.begin(), found_ptr)           \
+            );                                                                 \
+            return *start_val;                                                 \
+        }                                                                      \
+        constexpr name_type to_string() {                                      \
+            const auto* found_ptr = std::find_if(                              \
+                m_values.begin(),                                              \
+                m_values.end(),                                                \
+                [this](auto t_iter_val) { return t_iter_val == m_value; }      \
+            );                                                                 \
+            auto start_name = m_names.begin();                                 \
+            std::advance(                                                      \
+                start_name, std::distance(m_values.begin(), found_ptr)         \
+            );                                                                 \
+            return *start_name;                                                \
+        }                                                                      \
+    };                                                                         \
+    constexpr enum_name operator+(const enum_name::m_enumeration& t_val) {                \
+        return enum_name(t_val);                                               \
+    }                                                                          \
+// NOLINTEND(cppcoreguidelines-macro-usage)
 #endif