+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 #[Unreleased](https://github.com/bwhitchurch/CubeTimer/compare/0.1.1...HEAD)
 
+## New Features
+
+- feat: reflective enums through preprocessor shenanigans. [`671371a`](https://github.com/bwhitchurch/CubeTimer/commit/671371a80af9226e1778da5d8a71b2970a5b6d40)
+
 ## Miscellaneous
 
-- Enum things [`6575b6f`](https://github.com/bwhitchurch/CubeTimer/commit/6575b6f0c5b04ebe01838fb01754597988a7ae30)
+- Enum things [`8320a5e`](https://github.com/bwhitchurch/CubeTimer/commit/8320a5ee7d8ffa4e60959c8de8ed9dcadecc7703)
 #[0.1.1](https://github.com/bwhitchurch/CubeTimer/compare/0.1.0...0.1.1)
 
 ## New Features